tests: rename file in e2e test
