+++ tests/e2e/mv3-vite-vanilla-content-script-hmr/src1/main.ts
@@ -1,5 +1,5 @@
 import './style.css'
-import { header } from './content'
+import { header } from './header'
 
 const app = document.createElement('div')
 app.id = 'app'
