+++ tests/e2e/mv3-vite-vanilla-content-script-hmr/src2/header.ts
